changed data type of invoice item jmlKemasan to double

+++ src/java/com/enseval/ttss/model/Invoice.java
@@ -67,23 +67,23 @@ public class Invoice implements Serializable {
 
     String Currency = "IDR";
 
-    public Long getTotalNilai() {
-        Long total = listInvoiceItem.stream().mapToLong((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
+    public double getTotalNilai() {
+        double total = listInvoiceItem.stream().mapToDouble((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
         return total - ((total / 100) * this.getTax());
     }
 
-    public Long getTotalNilaiNoTax() {
-        Long total = listInvoiceItem.stream().mapToLong((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
+    public double getTotalNilaiNoTax() {
+        double total = listInvoiceItem.stream().mapToDouble((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
         return total;
     }
 
-    public Long getTaxValue() {
-        Long total = listInvoiceItem.stream().mapToLong((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
+    public double getTaxValue() {
+        double total = listInvoiceItem.stream().mapToDouble((InvoiceItem item) -> item.getHargaSatuan() * item.getJmlKemasan()).sum();
         return ((total / 100) * this.getTax());
     }
 
-    public Long getTotalTerbayar() {
-        Long total = listPelunasan.stream().mapToLong((Pelunasan item) -> item.getNilai() + item.getPotPPh() + item.getPotCN() + item.getPotAdm()).sum();
+    public double getTotalTerbayar() {
+        double total = listPelunasan.stream().mapToDouble((Pelunasan item) -> item.getNilai() + item.getPotPPh() + item.getPotCN() + item.getPotAdm()).sum();
         return total;
     }
