+++ src/java/com/enseval/ttss/vm/PopEditInvoiceVM.java
@@ -46,7 +46,7 @@ public class PopEditInvoiceVM {
     private Window winBuatInvoice;
     User userLogin;
     List<InvoiceItem> listInvoiceItem = new ArrayList<>();
-    Long totalHarga = 0L;
+    double totalHarga = 0;
 
     Invoice invoice;
 
@@ -220,7 +220,7 @@ public class PopEditInvoiceVM {
     @Command
     @NotifyChange({"totalHarga"})
     public void doCount() {
-        this.totalHarga = 0L;
+        this.totalHarga = 0;
         for (InvoiceItem item : listInvoiceItem) {
             this.totalHarga += item.getHargaSatuan() * item.getJmlKemasan();
         }
@@ -251,13 +251,14 @@ public class PopEditInvoiceVM {
         this.listInvoiceItem = listInvoiceItem;
     }
 
-    public Long getTotalHarga() {
+    public double getTotalHarga() {
         return totalHarga;
     }
 
-    public void setTotalHarga(Long totalHarga) {
+    public void setTotalHarga(double totalHarga) {
         this.totalHarga = totalHarga;
     }
+
 
     public Invoice getInvoice() {
         return invoice;
